backend/pyproject.toml
[project]
name = "vibehub-backend"
version = "0.1.0"
description = "VibeHub - AI-powered full-stack app generation backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy",
    "aiosqlite",
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "deepagents",
    "e2b",
    "langchain",
    "langchain-anthropic",
    "langchain-openai",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "langchain-google-genai>=4.2.1",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["api", "db", "services", "agent", "sandbox", "templates"]
